refactor: 컴포넌트 하나 위치 변경
