+++ app/note/_feature/components/NoteListCard.tsx
